Move bulk deletion tests to a separate spec file
